mpf/requirements.txt
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
idna==2.10
lml==0.1.0
lxml==4.6.2
numpy==1.19.4
odfpy==1.4.1
pandas==1.1.5
pkg-resources==0.0.0
pyexcel-ezodf==0.3.4
pyexcel-io==0.6.3
pyexcel-ods==0.6.0
pyexcel-ods3==0.6.0
pyexcel-xls==0.6.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
six==1.15.0
urllib3==1.26.2
xlrd==1.2.0
xlwt==1.3.0
